requirements.txt
appnope==0.1.0
awscli==1.11.136
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
bleach==2.0.0
botocore==1.6.3
certifi==2017.4.17
colorama==0.3.7
configparser==3.5.0
decorator==4.0.11
docutils==0.14
entrypoints==0.2.3
enum34==1.1.6
functools32==3.2.3.post2
futures==3.1.1
html5lib==0.999999999
ipykernel==4.6.1
ipython==5.4.1
ipython-genutils==0.2.0
ipywidgets==6.0.0
Jinja2==2.9.6
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.1.0
jupyter-core==4.3.0
MarkupSafe==1.0
mistune==0.7.4
mpmath==0.19
nbconvert==5.2.1
nbformat==4.3.0
nose==1.3.7
notebook==5.0.0
pandas==0.20.3
pandocfilters==1.4.1
pathlib2==2.3.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
ptyprocess==0.5.2
pyasn1==0.3.2
Pygments==2.2.0
python-dateutil==2.6.1
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.3.0
rsa==3.4.2
s3transfer==0.1.10
scandir==1.5
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
sympy==1.1
terminado==0.6
testpath==0.3.1
tornado==4.5.1
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==2.0.0
